remove old chunk header parser
